[varinder] update

+++ Payroll_System_Swift/extension+Int.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 
-extension Int
+extension Float
 {
     func addHours() -> String
     {
